refactor(core): Split source implementation in smaller modules
